Add `BrowserDetection` model concern
